rename symbols package to symbol
