+++ cmd/frontend/graphqlbackend/search_suggestions.go
@@ -22,7 +22,7 @@ import (
 	"github.com/sourcegraph/sourcegraph/internal/search/result"
 	"github.com/sourcegraph/sourcegraph/internal/search/searchcontexts"
 	"github.com/sourcegraph/sourcegraph/internal/search/streaming"
-	"github.com/sourcegraph/sourcegraph/internal/search/symbols"
+	"github.com/sourcegraph/sourcegraph/internal/search/symbol"
 )
 
 const maxSearchSuggestions = 100
@@ -397,7 +397,7 @@ func (r *searchResolver) Suggestions(ctx context.Context, args *searchSuggestion
 		defer cancel()
 
 		fileMatches, _, err := streaming.CollectStream(func(stream streaming.Sender) error {
-			return symbols.SearchSymbols(ctx, &search.TextParameters{
+			return symbol.SearchSymbols(ctx, &search.TextParameters{
 				PatternInfo:  p,
 				RepoPromise:  (&search.RepoPromise{}).Resolve(resolved.RepoRevs),
 				Query:        r.Query,
