+++ internal/search/run/aggregator.go
@@ -13,7 +13,7 @@ import (
 	searchrepos "github.com/sourcegraph/sourcegraph/internal/search/repos"
 	"github.com/sourcegraph/sourcegraph/internal/search/result"
 	"github.com/sourcegraph/sourcegraph/internal/search/streaming"
-	"github.com/sourcegraph/sourcegraph/internal/search/symbols"
+	"github.com/sourcegraph/sourcegraph/internal/search/symbol"
 	"github.com/sourcegraph/sourcegraph/internal/trace"
 )
 
@@ -86,7 +86,7 @@ func (a *Aggregator) DoSymbolSearch(ctx context.Context, args *search.TextParame
 		tr.Finish()
 	}()
 
-	err = symbols.SearchSymbols(ctx, args, limit, a)
+	err = symbol.SearchSymbols(ctx, args, limit, a)
 	return errors.Wrap(err, "symbol search failed")
 }
