+++ internal/search/symbol/symbols.go
@@ -1,4 +1,4 @@
-package symbols
+package symbol
 
 import (
 	"context"
